Added universal income as a "responsibility" for guests
