+++ code/jobs/Jobs.cs
@@ -5,7 +5,7 @@ using Sandbox;
 namespace Cinema.Jobs;
 
 /// <summary>
-/// Traits & Ablities that jobs have
+/// Traits & Abilities that jobs have
 /// </summary>
 [Flags]
 public enum JobAbilities : ulong
@@ -58,6 +58,7 @@ public partial class JobDetails : BaseNetworkable
         {
             Name = "Usher",
             Abilities = JobAbilities.PickupGarbage,
+            Responsibilities = 0,
         }
     };
 }
@@ -71,6 +72,8 @@ public partial class PlayerJob : EntityComponent<Player>, ISingletonComponent
     public new string Name => JobDetails?.Name ?? "Jobless";
 
     public JobAbilities Abilities => JobDetails?.Abilities ?? 0;
+
+    public JobResponsibilities Responsibilities => JobDetails?.Responsibilities ?? 0;
 
     public bool HasAbility(JobAbilities ability) => Abilities.HasFlag(ability);
