+++ code/jobs/Responsibilities.cs
@@ -20,27 +20,3 @@ public partial class JobResponsibility : EntityComponent<Player>
 {
     public new virtual string Name => "Responsibility";
 }
-
-public partial class UniversalIncome : JobResponsibility
-{
-    public override string Name => "Universal Income";
-
-    private static int IncomeAmount => 50;
-    private static float IncomeInterval => 60f;
-
-    private TimeUntil TimeUntilNextIncome { get; set; }
-
-    protected override void OnActivate()
-    {
-        TimeUntilNextIncome = IncomeInterval;
-    }
-
-    [GameEvent.Tick.Server]
-    protected void Tick()
-    {
-        if (TimeUntilNextIncome > 0) return;
-
-        TimeUntilNextIncome = IncomeInterval;
-        Entity.AddMoney(IncomeAmount);
-    }
-}
